chore(quality): raise PHPStan to level 8 and fix the bugs it surfaced
Level 8 adds array value types, generics and null-safety checks. Beyond
documenting array shapes (a SubscriberRecord type alias, resolver
`array<string, mixed>` args, batch result lists), it found:

- routes.php chained ->name() onto ->middleware(config(...)), which
  returns the middleware *array* when `laragraph.graphiql.middleware` is
  null — registering routes then fatals.
- ScaffoldCommand::render() did not handle file_get_contents() failure,
  and `laragraph:scaffold --all` still required a model argument.
- ConnectionType called paginate() on an untyped object; it now fails
  with a clear message for unsupported builders or paginators.
- Tracing span durations could be floats despite being typed as int.
- ExportSchemaCommand cast mixed console options to string.

+++ src/Tracing/TracingCollector.php
@@ -88,7 +88,7 @@ final class TracingCollector
             return;
         }
 
-        $this->spans[$spanId]['duration'] = (hrtime(true) - $this->startNs) - $this->spans[$spanId]['startOffset'];
+        $this->spans[$spanId]['duration'] = (int) ((hrtime(true) - $this->startNs) - $this->spans[$spanId]['startOffset']);
     }
 
     /**
